+++ api_service/src/main/java/es/kohchiku_bayashi/e_commerce_teahouse/repository/OrderProviderRepository.java
@@ -8,7 +8,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -27,10 +26,10 @@ public interface OrderProviderRepository extends JpaRepository<OrderProvider, Lo
     List<OrderProvider> findByMonthAndYear(@Param("month") int month, @Param("year") int year);
     
     @Query("SELECT SUM(op.total) FROM OrderProvider op")
-    BigDecimal getTotalCost();
+    Double getTotalCost();
     
     @Query("SELECT SUM(op.total) FROM OrderProvider op WHERE MONTH(op.orderDate) = :month")
-    BigDecimal getTotalCostByMonth(@Param("month") int month);
+    Double getTotalCostByMonth(@Param("month") int month);
     
     @Query("SELECT p.name, SUM(op.total) FROM OrderProvider op JOIN op.provider p GROUP BY p.name ORDER BY SUM(op.total)")
     List<Object[]> getTotalCostByProvider();
